+++ packages/mochi-rest/src/schemas/metadata.ts
@@ -8,15 +8,15 @@ export enum CommandScope {
 
 export const CommandSchema = z.object({
   id: z.number(),
-  code: z.string().nonempty(),
-  discord_command: z.string().nonempty(),
-  telegram_command: z.string().nonempty(),
+  code: z.string().min(1),
+  discord_command: z.string().min(1),
+  telegram_command: z.string().min(1),
   discord_alias: z.string(),
   telegram_alias: z.string(),
   scope: z.nativeEnum(CommandScope), // All / Private only / Public only
   description: z.string(),
-  created_at: z.string().datetime().nonempty(),
-  updated_at: z.string().datetime().nonempty(),
+  created_at: z.string().datetime(),
+  updated_at: z.string().datetime(),
 });
 
 export const ListCommandSchema = z.array(CommandSchema);
@@ -25,10 +25,10 @@ export type Command = z.infer<typeof CommandSchema>;
 
 export const CopySchema = z.object({
   id: z.number().nonnegative(),
-  type: z.string().nonempty(),
+  type: z.string().min(1),
   description: z.object({
-    fact: z.array(z.string().nonempty()),
-    tip: z.array(z.string().nonempty()),
+    fact: z.array(z.string().min(1)),
+    tip: z.array(z.string().min(1)),
   }),
   created_at: z.string().datetime(),
   updated_at: z.string().datetime(),
@@ -38,11 +38,11 @@ export type Copy = z.infer<typeof CopySchema>;
 
 export const ChangelogSchema = z.object({
   product: z.enum(["Mochi", "Mochi Discord", "Mochi Telegram", ""]),
-  title: z.string().nonempty(),
-  content: z.string().nonempty(),
+  title: z.string().min(1),
+  content: z.string().min(1),
   github_url: z.string().url().or(z.string()),
   thumbnail_url: z.string().url().or(z.string()),
-  file_name: z.string().nonempty(),
+  file_name: z.string().min(1),
   created_at: z.string().datetime(),
   updated_at: z.string().datetime(),
   is_expired: z.boolean(),
@@ -52,11 +52,23 @@ export const ListChangelogSchema = z.array(ChangelogSchema);
 export type Changelog = z.infer<typeof ChangelogSchema>;
 
 const EmojiSchema = z.object({
-  code: z.string().nonempty(),
-  emoji: z.string().nonempty(),
+  code: z.string().min(1),
+  emoji: z.string().min(1),
   emoji_url: z.string().url().or(z.literal("")),
 });
 
 export const EmojiListSchema = z.array(EmojiSchema);
 
 export type Emoji = z.infer<typeof EmojiSchema>;
+
+const ThemeSchema = z.object({
+  id: z.number().min(1),
+  image: z.string(),
+  slug: z.string(),
+  name: z.string(),
+  created_at: z.string().datetime(),
+  updated_at: z.string().datetime(),
+});
+
+export const ListThemeSchema = z.array(ThemeSchema);
+export type Theme = z.infer<typeof ThemeSchema>;